+++ appcli/commands/task_cli.py
@@ -56,10 +56,17 @@ class TaskCli:
             help="Runs a specified application task.",
             context_settings=dict(ignore_unknown_options=True),
         )
+        @click.option(
+            "--detach",
+            "-d",
+            is_flag=True,
+            default=False,
+            help="Run the task in the background.",
+        )
         @click.argument("service_name", required=True, type=click.STRING)
         @click.argument("extra_args", nargs=-1, type=click.UNPROCESSED)
         @click.pass_context
-        def run(ctx, service_name, extra_args):
+        def run(ctx, detach, service_name, extra_args):
             cli_context: CliContext = ctx.obj
             cli_context.get_configuration_dir_state().verify_command_allowed(
                 AppcliCommand.TASK_RUN
@@ -69,7 +76,9 @@ class TaskCli:
                 service_name,
                 extra_args,
             )
-            result = self.orchestrator.task(ctx.obj, service_name, extra_args)
+            result = self.orchestrator.task(
+                cli_context, service_name, extra_args, detached=detach
+            )
             logger.info("Task service finished with code [%i]", result.returncode)
             sys.exit(result.returncode)